Start of skill tree
Added start of skill tree
- Has different paths of skills (Tank, Fighter, Mobility)
-- Each path has different kinds of abilites/passive effects

- two different skill types (at time of push)
-- Active skills (require button press to use, have a cooldown)
-- Passive skills (always active, can be purchased in skill menu multiple times for stronger effect.

-Added skill that decreases attack speed by 5% every purchase, up to a 50% reduction.

- Can not buy a skill until at least one point is put into the previous skill in that tree (e.g cannot buy grip walls until triple jump is owned)

Misc
- Rounded all the attack speed values to 2 decimal places.

+++ Assets/Scripts/Class Scripts/Enemy.cs
@@ -11,7 +11,7 @@ public class Enemy
 {
     public string enemyName;
     public float enemyDamage;
-    public float enemyHealth;
+    public double enemyHealth;
     public float attackSpeed;
     public float attackRange;
 }
